+++ bindings/matrix-sdk-ffi/src/platform/tracing.rs
@@ -19,7 +19,7 @@ use std::{
     sync::{Arc, Mutex, OnceLock},
 };
 
-use tracing::{callsite::DefaultCallsite, debug, error, field::FieldSet, Callsite};
+use tracing::{Callsite, callsite::DefaultCallsite, debug, error, field::FieldSet};
 use tracing_core::{identify_callsite, metadata::Kind as MetadataKind};
 
 /// Log an event.